requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
albucore==0.0.13
albumentations==1.4.10
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
bce-python-sdk==0.9.29
beautifulsoup4==4.13.3
black==24.10.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
datasets==3.3.1
decorator==5.1.1
dill==0.3.4
docstring-to-markdown==0.15
easydict==1.13
exceptiongroup==1.2.2
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.16.1
fire==0.7.0
flake8==7.1.1
Flask==3.1.0
flask-babel==4.0.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.16.0
gradio_client==1.7.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.12.2
mypy-extensions==1.0.0
narwhals==1.34.1
networkx==3.4.2
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.20.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
orjson==3.10.15
packaging==24.2
paddle2onnx==1.3.1
paddlefsl==1.1.0
paddlehub==2.4.0
paddlenlp==2.6.1
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
protobuf==3.20.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyclipper==1.3.0.post6
pycodestyle==2.12.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.1
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.1
rarfile==4.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.2
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
seqeval==1.2.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchsummary==1.5.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
ultralytics==8.3.75
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
visualdl==2.5.3
websockets==14.2
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
